requirements.txt
python-fasthtml
mysql.connector